+++ src/components/omniSidebarNav/getGuideDirectory.js
@@ -1,36 +1,37 @@
 import { useStaticQuery, graphql } from 'gatsby';
+import allGuides from './allGuidesQuery';
 
-const getAllGuidePages = () => {
-  const AllTheGuides = useStaticQuery(
-    graphql`
-      {
-        allGuides: allMdx(
-          filter: {
-            fileAbsolutePath: { ne: null }
-            fields: { guide_directory: { ne: null } }
-            frontmatter: { draft: { ne: true } }
-          }
-          sort: { fields: [fileAbsolutePath], order: ASC }
-        ) {
-          edges {
-            node {
-              fields {
-                slug
-                guide_directory
-              }
-              frontmatter {
-                title
-                subtitle
-              }
-            }
-          }
-        }
-      }
-    `,
-  );
+// const getAllGuidePages = () => {
+//   const AllTheGuides = useStaticQuery(
+//     graphql`
+//       {
+//         allGuides: allMdx(
+//           filter: {
+//             fileAbsolutePath: { ne: null }
+//             fields: { guide_directory: { ne: null } }
+//             frontmatter: { draft: { ne: true } }
+//           }
+//           sort: { fields: [fileAbsolutePath], order: ASC }
+//         ) {
+//           edges {
+//             node {
+//               fields {
+//                 slug
+//                 guide_directory
+//               }
+//               frontmatter {
+//                 title
+//                 subtitle
+//               }
+//             }
+//           }
+//         }
+//       }
+//     `,
+//   );
 
-  return AllTheGuides;
-};
+//   return AllTheGuides;
+// };
 
 function getChildrenForGuideDirectory(AllTheGuides, guideDirectory) {
   const ChildItems = [];
@@ -54,7 +55,7 @@ function getTitleForGuideDirectory(AllTheGuides, guideDirectory) {
 }
 
 const getGuideDirectory = (guideDirectory) => {
-  const AllGuides = getAllGuidePages();
+  const AllGuides = allGuides();
   return {
     link: '/' + guideDirectory,
     title: getTitleForGuideDirectory(AllGuides.allGuides.edges, guideDirectory),
